fix a ulong/long comparison that clang was flagging.

+++ mq.c
@@ -500,7 +500,9 @@ MessageQueue_send(MessageQueue *self, PyObject *args, PyObject *keywords) {
         goto error_return;
     }
 
-    if (user_msg.len > self->max_message_size) {
+    // self->max_message_size is a ulong while user_msg.len is a long. Casting the latter to
+    // unsigned long is safe because the length will never be negative.
+    if ((unsigned long)user_msg.len > self->max_message_size) {
         PyErr_Format(PyExc_ValueError,
             "The message length exceeds queue's max_message_size (%lu)",
             self->max_message_size);
